refactor(filter): move Filter components back up
