--- packages/components/src/Filter/Filter/index.ts
@@ -1,1 +1,0 @@
-export * from "./Filter"
